# internal: Improve the Iter performance of LockMap on empty maps