<!--
* New contributors are highly encouraged to read our
  [CONTRIBUTING](/CONTRIBUTING.md) documentation.
* Commit and PR titles should be prefixed with the general area of the pull request's change.

-->
### What does this PR do?
Improves `LockMap.Iter` performance for empty maps
```
goos: darwin
goarch: arm64
pkg: gopkg.in/DataDog/dd-trace-go.v1/internal
                 │ original.out │          withAtomic2.out           │
                 │    sec/op    │   sec/op     vs base               │
Iter/original-10   14.280n ± 1%   1.998n ± 0%  -86.00% (p=0.002 n=6)

                 │ original.out │        withAtomic2.out        │
                 │     B/op     │    B/op     vs base           │
Iter/original-10     0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=6) ¹
¹ all samples are equal

                 │ original.out │        withAtomic2.out        │
                 │  allocs/op   │ allocs/op   vs base           │
Iter/original-10     0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=6) ¹
¹ all samples are equal

```
<!--
* A brief description of the change being made with this pull request.
* If the description here cannot be expressed in a succinct form, consider
  opening multiple pull requests instead of a single one.
-->

### Motivation
Turns out doing RLock and RUnlock can still be slow especially when this code is in a HOT path. For most users this map will be empty. Therefore it's worth speeding up that case at the cost of a tiny bit of extra memory and a bit of complexity.
<!--
* What inspired you to submit this pull request?
* Link any related GitHub issues or PRs here.
* If this resolves a GitHub issue, include "Fixes #XXXX" to link the issue and auto-close it on merge.
-->

### Describe how to test/QA your changes
Added a thrash test that when run with the race detector should help make sure we don't miss anything (and it verifies that the count and len match at the end to make sure they don't get out of sync).
<!--
* Write here in detail or link to detailed instructions on how this change can
  be tested/QAd/validated, including any environment setup.
* Ideally your changes have automated unit and/or integration tests which are run in CI.
-->

### Reviewer's Checklist
<!--
* Authors can use this list as a reference to ensure that there are no problems
  during the review but the signing off is to be done by the reviewer(s).
-->

- [ ] Changed code has unit tests for its functionality.
- [ ] If this interacts with the agent in a new way, a system test has been added.